+++ packages/common/src/core/traits/IWithViewSettings.ts
@@ -1,4 +1,5 @@
-import { IHubTimeline, IMetricDisplayConfig } from "../types";
+import { HubEntityHero } from "../../types";
+import { IHubTimeline, IMetricDisplayConfig, HubActionLink } from "../types";
 
 /**
  * Properties to be exclusively displayed on an entity's
@@ -38,4 +39,14 @@ export interface IWithViewSettings {
    * configurations for how to display metrics in the ui
    */
   metricDisplays?: IMetricDisplayConfig[];
+
+  /**
+   * what to show in the hero field, map/image
+   */
+  hero?: HubEntityHero;
+
+  /**
+   * array of actions for action links
+   */
+  heroActions?: HubActionLink[];
 }