+++ packages/common/src/initiatives/_internal/InitiativeSchema.ts
@@ -15,6 +15,10 @@ export const InitiativeSchema: IConfigurationSchema = {
   ...HubItemEntitySchema,
   properties: {
     ...HubItemEntitySchema.properties,
+    _groups: {
+      type: "array",
+      items: { type: "string" },
+    },
     status: {
       type: "string",
       default: HubEntityStatus.notStarted,