+++ packages/common/test/projects/_internal/ProjectUiSchemaEdit.test.ts
@@ -2,7 +2,7 @@ import { buildUiSchema } from "../../../src/projects/_internal/ProjectUiSchemaEd
 import { MOCK_CONTEXT } from "../../mocks/mock-auth";
 import * as getCategoryItemsModule from "../../../src/core/schemas/internal/getCategoryItems";
 import * as getFeaturedContentCatalogsModule from "../../../src/core/schemas/internal/getFeaturedContentCatalogs";
-import * as getAuthedImageUrlModule from "../../../src/core/schemas/internal/getAuthedImageUrl";
+import * as getAuthedImageUrlModule from "../../../src/core/_internal/getAuthedImageUrl";
 import * as getLocationExtentModule from "../../../src/core/schemas/internal/getLocationExtent";
 import * as getLocationOptionsModule from "../../../src/core/schemas/internal/getLocationOptions";
 import * as getTagItemsModule from "../../../src/core/schemas/internal/getTagItems";
@@ -91,11 +91,13 @@ describe("buildUiSchema: project edit", () => {
               scope: "/properties/description",
               type: "Control",
               options: {
-                control: "hub-field-input-input",
+                control: "hub-field-input-rich-text",
                 type: "textarea",
                 helperText: {
                   labelKey: "some.scope.fields.description.helperText",
                 },
+                toolbar:
+                  "heading,|,bold,italic,blockQuote,removeFormat,link,|,bulletedList,numberedList,alignment,outdent,indent,|,undo,redo",
               },
             },
             {
@@ -274,6 +276,45 @@ describe("buildUiSchema: project edit", () => {
             },
           ],
         },
+        {
+          type: "Section",
+          labelKey: "some.scope.sections.callToAction.label",
+          options: {
+            helperText: {
+              labelKey: "some.scope.sections.callToAction.helperText",
+            },
+          },
+          elements: [
+            {
+              scope: "/properties/view/properties/heroActions",
+              type: "Control",
+              options: {
+                control: "hub-composite-input-action-links",
+                type: "button",
+                catalogs: {},
+                facets: [
+                  {
+                    label: `{{some.scope.fields.callToAction.facets.type:translate}}`,
+                    key: "type",
+                    display: "multi-select",
+                    field: "type",
+                    options: [],
+                    operation: "OR",
+                    aggLimit: 100,
+                  },
+                  {
+                    label: `{{some.scope.fields.callToAction.facets.sharing:translate}}`,
+                    key: "access",
+                    display: "multi-select",
+                    field: "access",
+                    options: [],
+                    operation: "OR",
+                  },
+                ],
+              },
+            },
+          ],
+        },
       ],
     });
   });
@@ -363,11 +404,13 @@ describe("buildUiSchema: project edit", () => {
               scope: "/properties/description",
               type: "Control",
               options: {
-                control: "hub-field-input-input",
+                control: "hub-field-input-rich-text",
                 type: "textarea",
                 helperText: {
                   labelKey: "some.scope.fields.description.helperText",
                 },
+                toolbar:
+                  "heading,|,bold,italic,blockQuote,removeFormat,link,|,bulletedList,numberedList,alignment,outdent,indent,|,undo,redo",
               },
             },
             {
@@ -535,6 +578,45 @@ describe("buildUiSchema: project edit", () => {
                   {
                     label:
                       "{{some.scope.fields.featuredContent.facets.sharing:translate}}",
+                    key: "access",
+                    display: "multi-select",
+                    field: "access",
+                    options: [],
+                    operation: "OR",
+                  },
+                ],
+              },
+            },
+          ],
+        },
+        {
+          type: "Section",
+          labelKey: "some.scope.sections.callToAction.label",
+          options: {
+            helperText: {
+              labelKey: "some.scope.sections.callToAction.helperText",
+            },
+          },
+          elements: [
+            {
+              scope: "/properties/view/properties/heroActions",
+              type: "Control",
+              options: {
+                control: "hub-composite-input-action-links",
+                type: "button",
+                catalogs: {},
+                facets: [
+                  {
+                    label: `{{some.scope.fields.callToAction.facets.type:translate}}`,
+                    key: "type",
+                    display: "multi-select",
+                    field: "type",
+                    options: [],
+                    operation: "OR",
+                    aggLimit: 100,
+                  },
+                  {
+                    label: `{{some.scope.fields.callToAction.facets.sharing:translate}}`,
                     key: "access",
                     display: "multi-select",
                     field: "access",